requirements.txt
Faker
mysql-connector-python==8.3.0
apache-flink==1.17.2
elasticsearch==8.12.0
kafka-python